LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src/style-src-inline-style-attribute-blocked-expected.txt


PASS Inline style attribute should not be applied without 'unsafe-inline'
FAIL Should fire a securitypolicyviolation event assert_equals: expected "style-src" but got "style-src-attr"
Lorem ipsum
